Fix #13646 error sending sms from campagin
app/bundles/CampaignBundle/Executioner/Logger/EventLogger.php: We rasign the leads to lead event log objects to ensure they are managed by entity manager
before we attempt to persist event logs into database.

app/bundles/LeadBundle/EventListener/TimelineEventLogCampaignSubscriber.php: Similar to the EventLogger above ensure we work with managed entities at the
time we try persist it into database.

app/bundles/SmsBundle/Broadcast/BroadcastExecutioner.php: In referenced file we removed the broad try catch because in case of failure although script said
the job failed there was no way for anyone to know why it failed. Now the script terminates in case of exception with sms sending, but we also get full
traceback in logs that helps us resolve the issues that arise.

We might reintroduce the try-catch block back and just add a log entry in the future, if it turns out that removing the try-catch block is causing an issue.